refactor(config): isolate runtime resilience and watchdog policy interpretation
